requirements.txt
alphashape==1.3.1
collision==1.2.2
corgipath==0.1.0a1
cv_bridge==1.16.2
matplotlib==3.6.3
matplotlib==3.1.2
numpy==1.24.4
numpy==1.17.4
opencv_python==4.6.0.66
pandas==1.4.4
parmap==1.6.0
PyYAML==6.0
PyYAML==6.0.1
rospy==1.16.0
scikit_learn==1.1.2
scipy==1.11.4
seaborn==0.13.0
sensor_msgs==1.13.1
Shapely==2.0.2
std_msgs==0.0.1
tf==1.13.2
tf2_ros==0.7.7
torch==2.0.0+cu118
tqdm==4.64.1
trimesh==3.23.1
